@@ -1,19 +1,17 @@
-<template>
-  <div id="app">
-    <img src="./assets/logo.png">
-    <router-view/>
-  </div>
-</template>
 
+
+<template>
+  <img alt="Vue logo" src="./assets/logo.png" />
+  <router-view></router-view>
+</template>
 <script>
 export default {
   name: 'App'
 }
 </script>
-
 <style>
 #app {
-  font-family: 'Avenir', Helvetica, Arial, sans-serif;
+  font-family: Avenir, Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
